Fix: メニュー開閉時のコードをリファクタリング
