+++ app/javascript/controllers/menu.js
@@ -1,21 +1,21 @@
 // イベント間引き処理
-export const throttle = (fn, interval) => {
+export const throttle = (fn, arg, interval) => {
   let lastTime = Date.now() - interval
   return function() {
     if ((lastTime + interval) < Date.now()) {
       lastTime = Date.now()
-      fn()
+      fn(arg)
     }
   }
 }
 
 // メニュー開閉時のアニメーション
-export const animTiming = {
+const animTiming = {
   duration: 150,
   easing: 'linear'
 }
 
-export const closingAnimKeyframes = (content) => [
+const closingAnimKeyframes = (content) => [
   {
     height: content.offsetHeight + 'px',
     opacity: 1,
@@ -25,7 +25,7 @@ export const closingAnimKeyframes = (content) => [
   }
 ]
 
-export const openingAnimKeyframes = (content) => [
+const openingAnimKeyframes = (content) => [
   {
     height: 0,
     opacity: 0,
@@ -34,3 +34,19 @@ export const openingAnimKeyframes = (content) => [
     opacity: 1,
   }
 ]
+
+export function openMenu(content) {
+  menu.setAttribute('open', 'true')
+  content.classList.add('flex')
+  content.classList.remove('hidden')
+  content.animate(openingAnimKeyframes(content), animTiming)
+}
+
+export function closeMenu(content) {
+  const closingAnim = content.animate(closingAnimKeyframes(content), animTiming)
+  menu.removeAttribute('open')
+  closingAnim.onfinish = () => {
+    content.classList.add('hidden')
+    content.classList.remove('flex')
+  }
+}
